Refactor InputState and SoundManager
