+++ CMakeLists.txt
@@ -162,7 +162,9 @@ Set (FLARE_SOURCES
 	./src/RenderDeviceList.cpp
 	./src/SaveLoad.cpp
 	./src/SDL_gfxBlitFunc.c
+	./src/SDLInputState.cpp
 	./src/SDLSoftwareRenderDevice.cpp
+	./src/SDLSoundManager.cpp
 	./src/SDLHardwareRenderDevice.cpp
  	./src/SDLFontEngine.cpp
 	./src/Settings.cpp
@@ -170,7 +172,6 @@ Set (FLARE_SOURCES
 	./src/SharedResources.cpp
 	./src/StatBlock.cpp
 	./src/Stats.cpp
-	./src/SoundManager.cpp
 	./src/TileSet.cpp
 	./src/TooltipData.cpp
 	./src/Utils.cpp
@@ -265,8 +266,9 @@ Set (FLARE_HEADERS
 	./src/QuestLog.h
 	./src/RenderDevice.h
 	./src/RenderDeviceList.h
-	./src/SDL_gfxBlitFunc.c
+	./src/SDLInputState.h
 	./src/SDLSoftwareRenderDevice.h
+	./src/SDLSoundManager.h
 	./src/SDLHardwareRenderDevice.h
 	./src/SDLFontEngine.h
 	./src/Settings.h
