+++ src/InputState.h
@@ -78,21 +78,22 @@ public:
 	std::string mouse_button[7];
 
 	InputState(void);
-	~InputState();
+	virtual ~InputState() {};
 
-	void defaultQwertyKeyBindings();
 	void defaultJoystickBindings();
 	void loadKeyBindings();
 	void saveKeyBindings();
-	void handle(bool dump_event);
 	void resetScroll();
 	void lockActionBar();
 	void unlockActionBar();
 	void setKeybindNames();
-	void hideCursor();
-	void showCursor();
-	std::string getJoystickName(int index);
-	std::string getKeyName(int key);
+
+	virtual void handle(bool dump_event);
+	virtual void defaultQwertyKeyBindings() = 0;
+	virtual void hideCursor() = 0;
+	virtual void showCursor() = 0;
+	virtual std::string getJoystickName(int index) = 0;
+	virtual std::string getKeyName(int key) = 0;
 
 	bool pressing[key_count];
 	bool lock[key_count];
@@ -112,7 +113,7 @@ public:
 	bool window_restored;
 	bool window_resized;
 
-private:
+protected:
 	bool un_press[key_count];
 	Point current_touch;
 };
