+++ src/RenderDeviceList.cpp
@@ -24,6 +24,8 @@ FLARE.  If not, see http://www.gnu.org/licenses/
 #include "SDLHardwareRenderDevice.h"
 
 #include "SDLFontEngine.h"
+#include "SDLSoundManager.h"
+#include "SDLInputState.h"
 
 RenderDevice* getRenderDevice(std::string name) {
 #ifdef __ANDROID__
@@ -47,4 +49,12 @@ RenderDevice* getRenderDevice(std::string name) {
 
 FontEngine* getFontEngine() {
 	return new SDLFontEngine();
+}
+
+SoundManager* getSoundManager() {
+	return new SDLSoundManager();
+}
+
+InputState* getInputManager() {
+	return new SDLInputState();
 }
